HideGUI and Animate Entity - The GUI will now be hidden when user controls the camera and The playermodel will now animate through the recording -Hide the gui when user clicks on screen/away from gui -Player is now animated with walk, run and jump animations (Still VERY Limited)
What does this change bring?
This change allows users to hide the gui while they are controlling the
camera or simply holding right click outside of the gui. Player
recordings are now animated! However, it is severly limited to one walk
and one idle animation. This means that the player can now see
themselves animated while watching a recording.